backend/pyproject.toml
[project]
name = "app"
version = "2.1.31+pb-PE5whDilNl3pvCFn"
description = "Riff App"
authors = [{ name = "Databutton", email = "hi@databutton.com" }]

requires-python = ">=3.11,<3.12"
dependencies = []

[dependency-groups]
base = [
  "databutton==0.39.0",
  "uvicorn[standard]>=0.34.0",
  "fastapi>=0.115.7",
  "pydantic>=2.10.5",
  "httpx>=0.28.1",
  "python-multipart>=0.0.9",
  "pyjwt>=2.10.1",
  "cryptography>=44.0.0",
  "asyncpg>=0.30.0",
  "dotenv>=0.9.9",
]
dev = ["asyncpg-stubs>=0.30.2", "pytest>=8.4.2"]
app = ["openai", "beautifulsoup4", "requests", "email-validator"]

[tool.pytest.ini_options]
pythonpath = ["."]
